visualization-service/requirements.txt
paho-mqtt==2.1.0
influxdb-client==1.45.0
